services_proyect/requirements.txt
google-cloud-storage
kaggle==1.7.4.2
flask
pandas==2.2.3
protobuf==6.30.1
pymssql==2.3.2
pyodbc==5.2.0
SQLAlchemy==2.0.38
